Converted Gulpfile to ES6
