+++ cms/project_template/package.json
@@ -2,14 +2,16 @@
   "name": "{{ project_name }}",
   "main": "Gulpfile.js",
   "dependencies": {
-    "gulp": "^3.8.11"
+    "gulp": "^3.9.0"
   },
   "devDependencies": {
+    "babel": "^5.5.6",
     "browser-sync": "^2.2.1",
     "gulp-autoprefixer": "^2.1.0",
     "gulp-bower": "0.0.10",
     "gulp-changed": "^1.1.1",
     "gulp-if": "^1.2.5",
+    "gulp-load-plugins": "^0.10.0",
     "gulp-pxtorem": "^1.0.1",
     "gulp-rename": "^1.2.0",
     "gulp-sass": "^1.3.3",
